+++ config/test.exs
@@ -1,14 +1,11 @@
 use Mix.Config
 
-alias Commanded.EventStore.Adapters.InMemory
-alias Commanded.Serialization.JsonSerializer
-
 config :logger, :console, level: :warn, format: "[$level] $message\n"
 
 config :ex_unit,
   capture_log: true,
-  assert_receive_timeout: 200
+  assert_receive_timeout: 1_000
 
 config :commanded,
   dispatch_consistency_timeout: 100,
-  event_store_adapter: InMemory
+  event_store_adapter: Commanded.EventStore.Adapters.InMemory