+++ test/event/handle_event_test.exs
@@ -5,11 +5,11 @@ defmodule Commanded.Event.HandleEventTest do
   import Commanded.Assertions.EventAssertions
 
   alias Commanded.EventStore
-  alias Commanded.Event.{AppendingEventHandler,UninterestingEvent}
+  alias Commanded.Event.{AppendingEventHandler, UninterestingEvent}
   alias Commanded.Helpers.EventFactory
-  alias Commanded.Helpers.{ProcessHelper,Wait}
+  alias Commanded.Helpers.{ProcessHelper, Wait}
   alias Commanded.ExampleDomain.BankAccount.AccountBalanceHandler
-  alias Commanded.ExampleDomain.BankAccount.Events.{BankAccountOpened,MoneyDeposited}
+  alias Commanded.ExampleDomain.BankAccount.Events.{BankAccountOpened, MoneyDeposited}
 
   describe "balance handler" do
     setup do
@@ -19,9 +19,9 @@ defmodule Commanded.Event.HandleEventTest do
         assert AccountBalanceHandler.subscribed?()
       end)
 
-      on_exit fn ->
+      on_exit(fn ->
         ProcessHelper.shutdown(handler)
-      end
+      end)
 
       [handler: handler]
     end
@@ -31,12 +31,13 @@ defmodule Commanded.Event.HandleEventTest do
         %BankAccountOpened{account_number: "ACC123", initial_balance: 1_000},
         %MoneyDeposited{amount: 50, balance: 1_050}
       ]
+
       recorded_events = EventFactory.map_to_recorded_events(events)
 
       send(handler, {:events, recorded_events})
 
       Wait.until(fn ->
-        assert AccountBalanceHandler.current_balance == 1_050
+        assert AccountBalanceHandler.current_balance() == 1_050
       end)
     end
 
@@ -51,44 +52,53 @@ defmodule Commanded.Event.HandleEventTest do
         %MoneyDeposited{amount: 50, balance: 1_050},
         %UninterestingEvent{}
       ]
+
       recorded_events = EventFactory.map_to_recorded_events(events)
 
       send(handler, {:events, recorded_events})
 
       Wait.until(fn ->
-        assert AccountBalanceHandler.current_balance == 1_050
+        assert AccountBalanceHandler.current_balance() == 1_050
       end)
     end
   end
 
+  defp to_event_data(events) do
+    Commanded.Event.Mapper.map_to_event_data(events,
+      causation_id: UUID.uuid4(),
+      correlation_id: UUID.uuid4(),
+      metadata: %{}
+    )
+  end
+
   describe "appending handler" do
     setup do
-      on_exit fn ->
+      on_exit(fn ->
         ProcessHelper.shutdown(AppendingEventHandler)
-      end
+      end)
     end
 
     test "should ignore events created before the event handler's subscription when starting from `:current`" do
-      stream_uuid = UUID.uuid4
+      stream_uuid = UUID.uuid4()
       initial_events = [%BankAccountOpened{account_number: "ACC123", initial_balance: 1_000}]
       new_events = [%MoneyDeposited{amount: 50, balance: 1_050}]
 
-      {:ok, 1} = EventStore.append_to_stream(stream_uuid, 0, Commanded.Event.Mapper.map_to_event_data(initial_events, UUID.uuid4(), UUID.uuid4(), %{}))
+      :ok = EventStore.append_to_stream(stream_uuid, 0, to_event_data(initial_events))
 
-      wait_for_event BankAccountOpened
+      wait_for_event(BankAccountOpened)
 
       {:ok, handler} = AppendingEventHandler.start_link(start_from: :current)
 
       assert GenServer.call(handler, :last_seen_event) == nil
 
-      {:ok, 2} = EventStore.append_to_stream(stream_uuid, 1, Commanded.Event.Mapper.map_to_event_data(new_events, UUID.uuid4(), UUID.uuid4(), %{}))
+      :ok = EventStore.append_to_stream(stream_uuid, 1, to_event_data(new_events))
 
-      wait_for_event MoneyDeposited
+      wait_for_event(MoneyDeposited)
 
       Wait.until(fn ->
         assert AppendingEventHandler.received_events() == new_events
 
-        [ metadata ] = AppendingEventHandler.received_metadata()
+        [metadata] = AppendingEventHandler.received_metadata()
 
         assert Map.get(metadata, :event_number) == 2
         assert Map.get(metadata, :stream_id) == stream_uuid
@@ -97,20 +107,20 @@ defmodule Commanded.Event.HandleEventTest do
 
         assert GenServer.call(handler, :last_seen_event) == 2
       end)
-  	end
+    end
 
     test "should receive events created before the event handler's subscription when starting from `:origin`" do
-      stream_uuid = UUID.uuid4
+      stream_uuid = UUID.uuid4()
       initial_events = [%BankAccountOpened{account_number: "ACC123", initial_balance: 1_000}]
       new_events = [%MoneyDeposited{amount: 50, balance: 1_050}]
 
-      {:ok, 1} = EventStore.append_to_stream(stream_uuid, 0, Commanded.Event.Mapper.map_to_event_data(initial_events, UUID.uuid4(), UUID.uuid4(), %{}))
+      :ok = EventStore.append_to_stream(stream_uuid, 0, to_event_data(initial_events))
 
       {:ok, _handler} = AppendingEventHandler.start_link(start_from: :origin)
 
-      {:ok, 2} = EventStore.append_to_stream(stream_uuid, 1, Commanded.Event.Mapper.map_to_event_data(new_events, UUID.uuid4(), UUID.uuid4(), %{}))
+      :ok = EventStore.append_to_stream(stream_uuid, 1, to_event_data(new_events))
 
-      wait_for_event MoneyDeposited
+      wait_for_event(MoneyDeposited)
 
       Wait.until(fn ->
         assert AppendingEventHandler.received_events() == initial_events ++ new_events
@@ -125,15 +135,16 @@ defmodule Commanded.Event.HandleEventTest do
           assert %NaiveDateTime{} = Map.get(metadata, :created_at)
         end)
       end)
-  	end
+    end
 
-  	test "should ignore already seen events" do
+    test "should ignore already seen events" do
       {:ok, handler} = AppendingEventHandler.start_link()
 
       events = [
         %BankAccountOpened{account_number: "ACC123", initial_balance: 1_000},
         %MoneyDeposited{amount: 50, balance: 1_050}
       ]
+
       recorded_events = EventFactory.map_to_recorded_events(events)
 
       Wait.until(fn ->
@@ -150,7 +161,7 @@ defmodule Commanded.Event.HandleEventTest do
         assert AppendingEventHandler.received_events() == events
         assert pluck(AppendingEventHandler.received_metadata(), :stream_version) == [1, 2]
       end)
-  	end
+    end
   end
 
   describe "event handler name" do
@@ -181,19 +192,19 @@ defmodule Commanded.Event.HandleEventTest do
 
   test "should ensure an event handler name is provided" do
     assert_raise RuntimeError, "UnnamedEventHandler expects `:name` to be given", fn ->
-      Code.eval_string """
+      Code.eval_string("""
         defmodule UnnamedEventHandler do
           use Commanded.Event.Handler
         end
-      """
+      """)
     end
   end
 
   test "should allow using event handler module as name" do
-    Code.eval_string """
+    Code.eval_string("""
       defmodule EventHandler do
         use Commanded.Event.Handler, name: __MODULE__
       end
-    """
+    """)
   end
 end